using common mock for DB.test.ts
